// src/api/models/counsel-card-living-information-res.ts
/* tslint:disable */
/* eslint-disable */
/**
 * OpenAPI definition
 * No description provided (generated by Openapi Generator https://github.com/openapitools/openapi-generator)
 *
 * The version of the OpenAPI document: v0
 * 
 *
 * NOTE: This class is auto generated by OpenAPI Generator (https://openapi-generator.tech).
 * https://openapi-generator.tech
 * Do not edit the class manually.
 */


// May contain unused imports in some cases
// @ts-ignore
import type { DrinkingDTO } from './drinking-dto';
// May contain unused imports in some cases
// @ts-ignore
import type { ExerciseDTO } from './exercise-dto';
// May contain unused imports in some cases
// @ts-ignore
import type { MedicationManagementDTO } from './medication-management-dto';
// May contain unused imports in some cases
// @ts-ignore
import type { NutritionDTO } from './nutrition-dto';
// May contain unused imports in some cases
// @ts-ignore
import type { SmokingDTO } from './smoking-dto';

/**
 * 
 * @export
 * @interface CounselCardLivingInformationRes
 */
export interface CounselCardLivingInformationRes {
    /**
     * 
     * @type {DrinkingDTO}
     * @memberof CounselCardLivingInformationRes
     */
    'drinking'?: DrinkingDTO;
    /**
     * 
     * @type {ExerciseDTO}
     * @memberof CounselCardLivingInformationRes
     */
    'exercise'?: ExerciseDTO;
    /**
     * 
     * @type {MedicationManagementDTO}
     * @memberof CounselCardLivingInformationRes
     */
    'medicationManagement'?: MedicationManagementDTO;
    /**
     * 
     * @type {NutritionDTO}
     * @memberof CounselCardLivingInformationRes
     */
    'nutrition'?: NutritionDTO;
    /**
     * 
     * @type {SmokingDTO}
     * @memberof CounselCardLivingInformationRes
     */
    'smoking'?: SmokingDTO;
}